move schema to models
