+++ packages/npm/jns42-generator/src/generators/mocks-ts.ts
@@ -4,7 +4,7 @@ import {
   isOneOfSchemaModel,
   isSingleTypeSchemaModel,
   isTypeSchemaModel,
-} from "../schema/index.js";
+} from "../models/index.js";
 import {
   NestedText,
   banner,
