+++ packages/npm/jns42-generator/src/generators/types-ts.ts
@@ -5,7 +5,7 @@ import {
   isOneOfSchemaModel,
   isSingleTypeSchemaModel,
   isTypeSchemaModel,
-} from "../schema/index.js";
+} from "../models/index.js";
 import {
   NestedText,
   banner,
