+++ packages/npm/jns42-generator/src/models/model.ts
