+++ packages/npm/jns42-generator/src/schema-transforms/resolve-all-of.ts
@@ -4,7 +4,7 @@ import {
   intersectionMergeTypes,
   isAllOfSchemaModel,
   isSingleTypeSchemaModel,
-} from "../schema/index.js";
+} from "../models/index.js";
 import {
   booleanMergeAnd,
   booleanMergeOr,
