+++ packages/npm/jns42-generator/src/schema-transforms/resolve-if-then-else.test.ts
@@ -1,6 +1,6 @@
 import assert from "node:assert/strict";
 import test from "node:test";
-import { SchemaArena } from "../schema/arena.js";
+import { SchemaArena } from "../models/arena.js";
 import { normalizeObject } from "../utils/index.js";
 import { resolveIfThenElse } from "./resolve-if-then-else.js";
