--- packages/npm/jns42-generator/src/schema/index.ts
@@ -1,4 +1,0 @@
-export * from "./arena.js";
-export * from "./merge.js";
-export * from "./model.js";
-export * from "./selectors.js";
